fixed offset/scale for flash and group vars

+++ GOES/GOES-17/GLM/src/glm_data.h
@@ -80,7 +80,7 @@ typedef struct GLM_EVENT
     unsigned int time_offset;
     float lat;
     float lon;
-    unsigned int energy;
+    float energy;
     unsigned int parent_group_id;
 } GLM_EVENT_T;
 
@@ -90,8 +90,8 @@ typedef struct GLM_GROUP
     unsigned int time_offset;
     float lat;
     float lon;
-    unsigned int energy;
-    unsigned int area;
+    float energy;
+    float area;
     unsigned int parent_flash_id;
     short quality_flag;
 } GLM_GROUP_T;
